Server: Added Postgresql dependency, updated libs to latest x64 versions, switched build to x64

+++ Server/Crypto/Rsa.cpp
@@ -36,7 +36,7 @@ std::vector<uint8_t> Rsa::Encrypt(std::vector<uint8_t> const& data, unsigned cha
 
     int(*encryptFnc)(int, unsigned char const*, unsigned char*, RSA*, int) = isPublic ? RSA_public_encrypt : RSA_private_encrypt;
 
-    int encryptedLength = encryptFnc(data.size(), data.data(), encrypted, rsa, RSA_PKCS1_PADDING);
+    int encryptedLength = encryptFnc((int)data.size(), data.data(), encrypted, rsa, RSA_PKCS1_PADDING);
     if (encryptedLength == -1)
         return data;
 
@@ -59,12 +59,12 @@ std::vector<uint8_t> Rsa::Decrypt(std::vector<uint8_t> const& data, unsigned cha
 
     int(*decryptFnc)(int, unsigned char const*, unsigned char*, RSA*, int) = isPublic ? RSA_public_decrypt : RSA_private_decrypt;
 
-    int decryptedLength = decryptFnc(data.size(), data.data(), decrypted, rsa, RSA_PKCS1_PADDING);
+    int decryptedLength = decryptFnc((int)data.size(), data.data(), decrypted, rsa, RSA_PKCS1_PADDING);
     if (decryptedLength == -1)
         return data;
 
     std::vector<uint8_t> value(decryptedLength);
     std::memcpy(&value[0], decrypted, decryptedLength);
-
+    
     return value;
 }
